requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi_cors==0.0.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
